Split tls config tests
